error solved

+++ src/utils/cloudinary.js
@@ -9,15 +9,18 @@ cloudinary.config({
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if(!localFilePath) return null;
-        // upload the file on cloudinary
+        if (!localFilePath) return null;
+
         const response = await cloudinary.uploader.upload(localFilePath, {
-            resource_type: "auto"
-        })
-        // file has been uploaded successfully
+            resource_type: "auto",
+        });
+
         console.log("file is uploaded on cloudinary!", response.url);
+        return response; // important
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+        if (localFilePath) {
+            fs.unlinkSync(localFilePath);
+        }
         return null;
     }
 }
